+++ src/value.h
@@ -14,7 +14,7 @@ typedef struct ValueArray {
 void value_array_init(ValueArray *value_array);
 void value_array_free(ValueArray *value_array);
 void value_array_write(ValueArray *value_array, Value value);
-Value value_get(ValueArray *array, uint8_t i);
+Value value_get(ValueArray *array, int i);
 void value_print(Value value);
 
 #endif
